Move social links to file
